+++ hippodamus-tests/test/dd/kms/hippodamus/LoggingTest.java
@@ -1,17 +1,23 @@
 package dd.kms.hippodamus;
 
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Set;
+
+import javax.annotation.Nullable;
+
 import dd.kms.hippodamus.api.coordinator.Coordinators;
 import dd.kms.hippodamus.api.coordinator.ExecutionCoordinator;
 import dd.kms.hippodamus.api.coordinator.configuration.ExecutionCoordinatorBuilder;
 import dd.kms.hippodamus.api.exceptions.ExceptionalRunnable;
-import dd.kms.hippodamus.api.logging.LogLevel;
+import dd.kms.hippodamus.api.handles.Handle;
+import dd.kms.hippodamus.api.handles.TaskStage;
 import dd.kms.hippodamus.api.logging.Logger;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.MethodSource;
 
-import java.util.HashMap;
-import java.util.Map;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+import com.google.common.collect.ImmutableSet;
 
 /**
  * This test verifies that log messages are sent to the specified logger and that messages with log levels that are
@@ -21,49 +27,59 @@ class LoggingTest
 {
 	private static final int	NUM_TASKS	= 10;
 
-	@ParameterizedTest(name = "minimum log level: {0}")
-	@MethodSource("getLogLevelValues")
-	void testLogger(LogLevel minLogLevel) {
+	@Test
+	void testLogger() {
 		TestLogger logger = new TestLogger();
 		ExecutionCoordinatorBuilder coordinatorBuilder = Coordinators.configureExecutionCoordinator()
-			.logger(logger)
-			.minimumLogLevel(minLogLevel);
+			.logger(logger);
 		try (ExecutionCoordinator coordinator = coordinatorBuilder.build()) {
 			ExceptionalRunnable<RuntimeException> task = () -> {};
 			for (int i = 0; i < NUM_TASKS; i++) {
 				coordinator.execute(task);
 			}
 		}
-		Assertions.assertEquals(0, logger.getCounter(LogLevel.INTERNAL_ERROR), "An internal error occurred");
+		Assertions.assertEquals(0, logger.getNumberOfErrors(), "An internal error occurred");
 
-		int expectedNumStateChangeLogMessages = minLogLevel.compareTo(LogLevel.STATE) >= 0
-				? NUM_TASKS*5	// logged 5 state changes per task: SUBMITTED, EXECUTING, FINISHED, notification about result, and TERMINATED
-				: 0;			// state changes not logged
-		Assertions.assertEquals(expectedNumStateChangeLogMessages, logger.getCounter(LogLevel.STATE), "Wrong number of state change log messages");
-	}
-
-	static Object getLogLevelValues() {
-		return LogLevel.values();
+		Set<TaskStage> expectedTaskStages = ImmutableSet.of(TaskStage.READY, TaskStage.SUBMITTED, TaskStage.EXECUTING, TaskStage.FINISHED, TaskStage.TERMINATED);
+		for (TaskStage taskStage : TaskStage.values()) {
+			boolean stageExpected = expectedTaskStages.contains(taskStage);
+			int expectedNumOccurrences = stageExpected ? NUM_TASKS : 0;
+			Assertions.assertEquals(expectedNumOccurrences, logger.getNumberOfChangesTo(taskStage), "Wrong number of state change log messages");
+		}
 	}
 
 	private static class TestLogger implements Logger
 	{
-		private final Map<LogLevel, Integer>	logLevelCounters	= new HashMap<>();
+		private final Map<TaskStage, Integer>	taskStageCounters	= new HashMap<>();
+		private int 							numErrors			= 0;
 
 		TestLogger() {
-			for (LogLevel logLevel : LogLevel.values()) {
-				logLevelCounters.put(logLevel, 0);
+			for (TaskStage taskStage : TaskStage.values()) {
+				taskStageCounters.put(taskStage, 0);
 			}
 		}
 
-		@Override
-		public void log(LogLevel logLevel, String taskName, String message) {
-			int counter = logLevelCounters.get(logLevel);
-			logLevelCounters.put(logLevel, counter + 1);
+		public void log(@Nullable Handle handle, String message) {
+			/* do nothing */
 		}
 
-		int getCounter(LogLevel logLevel) {
-			return logLevelCounters.get(logLevel);
+		@Override
+		public void logStateChange(Handle handle, TaskStage taskStage) {
+			int counter = taskStageCounters.get(taskStage);
+			taskStageCounters.put(taskStage, counter + 1);
+		}
+
+		@Override
+		public void logError(@Nullable Handle handle, String error, @Nullable Throwable cause) {
+			numErrors++;
+		}
+
+		int getNumberOfChangesTo(TaskStage taskStage) {
+			return taskStageCounters.get(taskStage);
+		}
+
+		int getNumberOfErrors() {
+			return numErrors;
 		}
 	}
 }