+++ hippodamus/src/dd/kms/hippodamus/impl/coordinator/configuration/ExecutionCoordinatorBuilderImpl.java
@@ -1,14 +1,13 @@
 package dd.kms.hippodamus.impl.coordinator.configuration;
 
+import java.util.Map;
+
 import dd.kms.hippodamus.api.coordinator.ExecutionCoordinator;
 import dd.kms.hippodamus.api.coordinator.TaskType;
 import dd.kms.hippodamus.api.coordinator.configuration.ExecutionCoordinatorBuilder;
-import dd.kms.hippodamus.api.logging.LogLevel;
 import dd.kms.hippodamus.api.logging.Logger;
 import dd.kms.hippodamus.impl.coordinator.ExecutionCoordinatorImpl;
 import dd.kms.hippodamus.impl.execution.ExecutorServiceWrapper;
-
-import java.util.Map;
 
 public class ExecutionCoordinatorBuilderImpl extends CoordinatorBuilderBase<ExecutionCoordinatorBuilder, ExecutionCoordinator> implements ExecutionCoordinatorBuilder
 {
@@ -18,7 +17,7 @@ public class ExecutionCoordinatorBuilderImpl extends CoordinatorBuilderBase<Exec
 	}
 
 	@Override
-	ExecutionCoordinator createCoordinator(Map<TaskType, ExecutorServiceWrapper> executorServiceWrappersByTaskType, Logger logger, LogLevel minimumLogLevel, boolean verifyDependencies) {
-		return new ExecutionCoordinatorImpl(executorServiceWrappersByTaskType, logger, minimumLogLevel, verifyDependencies);
+	ExecutionCoordinator createCoordinator(Map<TaskType, ExecutorServiceWrapper> executorServiceWrappersByTaskType, Logger logger, boolean verifyDependencies) {
+		return new ExecutionCoordinatorImpl(executorServiceWrappersByTaskType, logger, verifyDependencies);
 	}
 }